camera_project/requirements.txt
# Project dependencies

opencv-python>=4.10.0
Pillow>=10.4.0
numpy>=1.26.4